requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.6.2
redis==7.2.0
django-celery-beat==2.8.1
amqp==5.3.1
billiard==4.2.4
kombu==5.6.2
vine==5.1.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==2.0.6
django-timezone-field==7.2.1
packaging==26.0
prompt-toolkit==3.0.52
python-crontab==3.3.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
wcwidth==0.6.0